INTDEV-864 Split Calculate command to command and engine parts
