+++ tools/data-handler/src/commands/validate.ts
@@ -52,7 +52,6 @@ const SHORT_TEXT_MAX_LENGTH = 80;
 
 import * as EmailValidator from 'email-validator';
 import { evaluateMacros } from '../macros/index.js';
-import { Calculate } from './calculate.js';
 const baseDir = dirname(fileURLToPath(import.meta.url));
 const subFoldersToValidate = ['.cards', 'cardRoot'];
 
@@ -547,7 +546,6 @@ export class Validate {
       } else {
         const errorMsg: string[] = [];
         const project = new Project(projectPath);
-        const calculate = new Calculate(project);
 
         // Then, validate that each 'contentSchema' children as well.
         const result = await this.readAndValidateContentFiles(
@@ -607,7 +605,7 @@ export class Validate {
                 project,
                 cardKey: card.key,
               },
-              calculate,
+              project.calculationEngine,
             );
           }
         }
