+++ tools/data-handler/src/macros/index.ts
@@ -34,8 +34,8 @@ import type {
 } from '../interfaces/macros.js';
 import type BaseMacro from './base-macro.js';
 import TaskQueue from './task-queue.js';
-import type { Calculate } from '../commands/index.js';
 import { ClingoError } from '@cyberismo/node-clingo';
+import type { CalculationEngine } from '../containers/project/calculation-engine.js';
 const CURLY_LEFT = '&#123;';
 const CURLY_RIGHT = '&#125;';
 const RAW_BLOCK_OPEN = '{{#raw}}';
@@ -126,7 +126,7 @@ export interface SimpleMacroConstructor {
  * Constructor for report macros
  */
 export interface ReportMacroConstructor {
-  new (tasks: TaskQueue, calculate: Calculate): BaseMacro;
+  new (tasks: TaskQueue, calculate: CalculationEngine): BaseMacro;
 }
 
 /**
@@ -186,7 +186,7 @@ export function registerMacros(
   instance: typeof Handlebars,
   context: MacroGenerationContext,
   tasks: TaskQueue,
-  calculate: Calculate,
+  calculate: CalculationEngine,
 ) {
   const macroInstances: BaseMacro[] = [];
   for (const macro of Object.keys(macros) as MacroName[]) {
@@ -232,7 +232,7 @@ export function registerEmptyMacros(instance: typeof Handlebars) {
 export async function evaluateMacros(
   content: string,
   context: MacroGenerationContext,
-  calculate: Calculate,
+  calculate: CalculationEngine,
 ) {
   const handlebars = Handlebars.create();
   const tasks = new TaskQueue();
