+++ tools/data-handler/test/command-handler-import.test.ts
@@ -11,7 +11,7 @@ import { fileURLToPath } from 'node:url';
 import { copyDir } from '../src/utils/file-utils.js';
 import { type CardsOptions, Cmd, Commands } from '../src/command-handler.js';
 import { Project } from '../src/containers/project.js';
-import { Calculate, Show } from '../src/commands/index.js';
+import { Show } from '../src/commands/index.js';
 
 // Create test artifacts in a temp folder.
 const baseDir = dirname(fileURLToPath(import.meta.url));
@@ -45,8 +45,7 @@ describe('import csv command', () => {
     const [key1, key2] = result.payload as string[];
 
     const project = new Project(decisionRecordsPath);
-    const calculate = new Calculate(project);
-    const show = new Show(project, calculate);
+    const show = new Show(project);
     const card1 = await show.showCardDetails(
       { metadata: true, content: true },
       key1,
@@ -84,8 +83,7 @@ describe('import csv command', () => {
 
     const createdKeys = result.payload as string[];
     const project = new Project(decisionRecordsPath);
-    const calculate = new Calculate(project);
-    const show = new Show(project, calculate);
+    const show = new Show(project);
 
     const parentCard = await show.showCardDetails(
       { metadata: true, content: true, children: true },
